requirements.txt
Django==1.6
MySQL-python==1.2.4
Unipath==1.0
django-bootstrap3==2.0.0
django-braces==1.2.2
django-debug-toolbar==0.11.0
django-extensions==1.2.5
inspector-panel==0.0.1
six==1.4.1
sqlparse==0.1.10
